requirements.txt
streamlit
gtts
pillow
streamlit-authenticator
